refactor : homeScreen 리팩토링
